requirements.txt
# requirements.txt
requests
mysql-connector-python
pytz